demystify/demystify_stack.py
from aws_cdk import (
    core as cdk,
)


class DemystifyStack(cdk.Stack):

    def __init__(self, scope: cdk.Construct, construct_id: str, **kwargs) -> None:
        super().__init__(scope, construct_id, **kwargs)
