requirements.txt
aws-cdk.core
aws-cdk.assertions
